separate server error source util to reduce bundle size
